+++ fast_rx/lib/fast_rx.dart
@@ -17,7 +17,6 @@ export 'src/rx/object/rx_set.dart';
 export 'src/rx/primitive/rx_bool.dart';
 export 'src/rx/primitive/rx_double.dart';
 export 'src/rx/primitive/rx_int.dart';
-export 'src/rx/primitive/rx_num.dart';
 export 'src/rx/primitive/rx_string.dart';
 
 // rx